+++ c/macos.c
@@ -21,37 +21,49 @@ static int skipvfs;
 
 /* Get information */
 const char *get_os_type(void) {
-	char *s, buf[LEN];
+	char *buf;
 	size_t len;
 	int mib[2];
 
 	mib[0] = CTL_KERN;
 	mib[1] = KERN_OSTYPE;
-	s = malloc(LEN);
-	len = sizeof(buf);
-	
-	if (sysctl(mib, 2, buf, &len, NULL, 0) == -1)
-		strncpy(s, "Darwin", len);
-	strncpy(s, buf, len);
+	buf = malloc(LEN);
+	len = LEN;
 
-	return s;
+	if (sysctl(mib, 2, buf, &len, NULL, 0) == -1) {
+		// At this point `len` is _most likely_ set to `0` (e.g. in case of ENOMEM).
+		// We copy our string and reset the length.
+		strncpy(buf, "Darwin", LEN);
+		len = LEN;
+	}
+
+	// Ensure null termination at the end.
+	buf[len-1] = '\0';
+
+	return buf;
 }
 
 const char *get_os_release(void) {
-	char *s, buf[LEN];
+	char *buf;
 	size_t len;
 	int mib[2];
 
 	mib[0] = CTL_KERN;
 	mib[1] = KERN_OSRELEASE;
-	s = malloc(LEN);
-	len = sizeof(buf);
+	buf = malloc(LEN);
+	len = LEN;
 
-	if (sysctl(mib, 2, buf, &len, NULL, 0) == -1)
-		strncpy(s, "Unknown", len);
-	strncpy(s, buf, len);
+	if (sysctl(mib, 2, buf, &len, NULL, 0) == -1) {
+		// At this point `len` is _most likely_ set to `0` (e.g. in case of ENOMEM).
+		// We copy our string and reset the length.
+		strncpy(buf, "Unknown", LEN);
+		len = LEN;
+	}
 
-	return s;
+	// Ensure null termination at the end.
+	buf[len-1] = '\0';
+
+	return buf;
 }
 
 unsigned int get_cpu_num(void) {
@@ -72,7 +84,7 @@ unsigned int get_cpu_num(void) {
 unsigned long get_cpu_speed(void) {
 	unsigned long speed;
 	size_t len;
-	
+
 	len = sizeof(speed);
 	sysctlbyname("hw.cpufrequency", &speed, &len, NULL, 0);
 	speed /= 1000000;
@@ -89,7 +101,7 @@ unsigned long get_proc_total(void) {
 	mib[2] = KERN_PROC_ALL;
 
 	sysctl(mib, 3, NULL, &len, NULL, 0);
-	
+
 	return len / sizeof(struct kinfo_proc);
 }
 
@@ -98,6 +110,7 @@ MemInfo get_mem_info(void) {
 	size_t len;
 	int mib[2];
 	vm_statistics_data_t vm_stat;
+	struct xsw_usage swap_info;
 	mach_msg_type_number_t count = HOST_VM_INFO_COUNT;
 	MemInfo mi;
 
@@ -108,17 +121,22 @@ MemInfo get_mem_info(void) {
 		sysctl(mib, 2, &size, &len, NULL, 0);
 		size /= 1024;
 	}
-	
+
+	mib[0] = CTL_VM;
+  	mib[1] = VM_SWAPUSAGE;
+	len = sizeof(swap_info);
+	sysctl(mib, 2 , &swap_info, &len, NULL, 0);
+
 	host_statistics(mach_host_self(), HOST_VM_INFO,
 				(host_info_t)&vm_stat, &count);
 
 	mi.total       = size;
-	mi.avail       = vm_stat.active_count * PAGE_SIZE / 1024;
-	mi.free        = vm_stat.free_count * PAGE_SIZE / 1024;
+	mi.avail       = (vm_stat.free_count + vm_stat.inactive_count) * PAGE_SIZE / 1024;
+	mi.free        = (vm_stat.free_count - vm_stat.speculative_count) * PAGE_SIZE / 1024;
 	mi.buffers     = 0;
 	mi.cached      = 0;
-	mi.swap_total  = 0;
-	mi.swap_free   = 0;
+	mi.swap_total  = swap_info.xsu_total / 1024;
+	mi.swap_free   = swap_info.xsu_avail / 1024;
 
 	return mi;
 }
@@ -275,4 +293,4 @@ char *makenetvfslist(void){
 }
 
 
-	
+