+++ lib.rs
@@ -11,16 +11,16 @@ use std::ffi;
 use std::fmt;
 use std::io::{self, BufRead, BufReader};
 use std::fs::{read_to_string, File};
-#[cfg(any(target_os = "windows", target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
+#[cfg(any(target_os = "windows", target_os = "macos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
 use std::os::raw::c_char;
 #[cfg(not(any(target_os = "windows", target_os = "linux")))]
 use std::os::raw::{c_double, c_int};
 
-#[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
+#[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
 use libc::sysctl;
-#[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
+#[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
 use std::mem::size_of_val;
-#[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
+#[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
 use std::ptr::null_mut;
 #[cfg(not(target_os = "windows"))]
 use libc::timeval;
@@ -32,20 +32,20 @@ use std::collections::HashMap;
 #[cfg(any(target_os = "solaris", target_os = "illumos"))]
 mod kstat;
 
-#[cfg(any(target_os = "macos", target_os="freebsd", target_os = "openbsd"))]
+#[cfg(any(target_os = "macos", target_os="freebsd", target_os = "openbsd", target_os = "netbsd"))]
 static OS_CTL_KERN: libc::c_int = 1;
-#[cfg(any(target_os = "macos", target_os="freebsd", target_os = "openbsd"))]
+#[cfg(any(target_os = "macos", target_os="freebsd", target_os = "openbsd", target_os = "netbsd"))]
 static OS_KERN_BOOTTIME: libc::c_int = 21;
 
 /// System load average value.
 #[repr(C)]
 #[derive(Debug)]
 pub struct LoadAvg {
-    /// Average load within one minite.
+    /// Average load within one minutes.
     pub one: f64,
-    /// Average load within five minites.
+    /// Average load within five minutes.
     pub five: f64,
-    /// Average load within fifteen minites.
+    /// Average load within fifteen minutes.
     pub fifteen: f64,
 }
 
@@ -202,7 +202,7 @@ pub struct LinuxOSReleaseInfo {
     /// Examples: "server", "embedded".
     pub variant_id: Option<String>,
 
-    /// HOME_URL= should refer to the of the operating system, or alternatively some homepage of
+    /// HOME_URL= should refer to the homepage of the operating system, or alternatively some homepage of
     /// the specific version of the operating system.
     ///
     /// These URLs are intended to be exposed in "About this system" UIs behind links with captions
@@ -352,27 +352,27 @@ impl From<Box<dyn std::error::Error>> for Error {
 extern "C" {
     #[cfg(any(target_os = "macos", target_os = "windows"))]
     fn get_os_type() -> *const i8;
-    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     fn get_os_release() -> *const i8;
 
-    #[cfg(all(not(any(target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd")), any(unix, windows)))]
+    #[cfg(all(not(any(target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")), any(unix, windows)))]
     fn get_cpu_num() -> u32;
-    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     fn get_cpu_speed() -> u64;
 
     #[cfg(target_os = "windows")]
     fn get_loadavg() -> LoadAvg;
-    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     fn get_proc_total() -> u64;
 
     #[cfg(any(target_os = "macos", target_os = "windows"))]
     fn get_mem_info() -> MemInfo;
-    #[cfg(any(target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     fn get_mem_info_bsd(mi: &mut MemInfo) -> i32;
 
     #[cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))]
     fn get_disk_info() -> DiskInfo;
-    #[cfg(any(target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     fn get_disk_info_bsd(di: &mut DiskInfo) -> i32;
 }
 
@@ -408,7 +408,11 @@ pub fn os_type() -> Result<String, Error> {
     {
         Ok("openbsd".to_string())
     }
-    #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows", target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(target_os = "netbsd")]
+    {
+        Ok("netbsd".to_string())
+    }
+    #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows", target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -424,7 +428,7 @@ pub fn os_release() -> Result<String, Error> {
         debug_assert!(s.pop() == Some('\n')); // pop '\n'
         Ok(s)
     }
-    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         unsafe {
             let rp = get_os_release() as *const c_char;
@@ -452,7 +456,7 @@ pub fn os_release() -> Result<String, Error> {
             Some(release) => Ok(release),
         }
     }
-    #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows", target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows", target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -525,7 +529,7 @@ fn parse_line_for_linux_os_release(line: &str) -> Option<(&str, String)> {
 ///
 /// Notice, it returns the logical cpu quantity.
 pub fn cpu_num() -> Result<u32, Error> {
-    #[cfg(any(target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "solaris", target_os = "illumos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let ret = unsafe { libc::sysconf(libc::_SC_NPROCESSORS_ONLN) };
         if ret < 1 || ret > std::u32::MAX as i64 {
@@ -534,7 +538,7 @@ pub fn cpu_num() -> Result<u32, Error> {
             Ok(ret as u32)
         }
     }
-    #[cfg(all(not(any(target_os = "solaris", target_os = "illumos", target_os="freebsd", target_os = "openbsd")), any(unix, windows)))]
+    #[cfg(all(not(any(target_os = "solaris", target_os = "illumos", target_os="freebsd", target_os = "openbsd", target_os = "netbsd")), any(unix, windows)))]
     {
         unsafe { Ok(get_cpu_num()) }
     }
@@ -578,7 +582,7 @@ pub fn cpu_speed() -> Result<u64, Error> {
     {
         unsafe { Ok(get_cpu_speed()) }
     }
-    #[cfg(any(target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let res: u64 = unsafe { get_cpu_speed() };
         match res {
@@ -586,7 +590,7 @@ pub fn cpu_speed() -> Result<u64, Error> {
             _ => Ok(res),
         }
     }
-    #[cfg(not(any(target_os = "solaris", target_os = "illumos", target_os = "linux", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(not(any(target_os = "solaris", target_os = "illumos", target_os = "linux", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -606,7 +610,7 @@ pub fn loadavg() -> Result<LoadAvg, Error> {
             fifteen: loads.next().unwrap(),
         })
     }
-    #[cfg(any(target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let mut l: [c_double; 3] = [0f64; 3];
         if unsafe { libc::getloadavg(l.as_mut_ptr(), l.len() as c_int) } < 3 {
@@ -623,7 +627,7 @@ pub fn loadavg() -> Result<LoadAvg, Error> {
     {
         Ok(unsafe { get_loadavg() })
     }
-    #[cfg(not(any(target_os = "linux", target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(not(any(target_os = "linux", target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -647,7 +651,7 @@ pub fn proc_total() -> Result<u64, Error> {
     {
         Ok(unsafe { get_proc_total() })
     }
-    #[cfg(any(target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let res: u64 = unsafe { get_proc_total() };
         match res {
@@ -655,7 +659,7 @@ pub fn proc_total() -> Result<u64, Error> {
             _ => Ok(res),
         }
     }
-    #[cfg(not(any(target_os = "linux", target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(not(any(target_os = "linux", target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -728,7 +732,7 @@ pub fn mem_info() -> Result<MemInfo, Error> {
     {
         Ok(unsafe { get_mem_info() })
     }
-    #[cfg(any(target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let mut mi:MemInfo = MemInfo {
             total: 0, free: 0, avail: 0, buffers: 0, cached: 0, swap_total: 0, swap_free: 0,
@@ -740,7 +744,7 @@ pub fn mem_info() -> Result<MemInfo, Error> {
             _ => Err(Error::Unknown),
         }
     }
-    #[cfg(not(any(target_os = "linux", target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(not(any(target_os = "linux", target_os = "solaris", target_os = "illumos", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -754,7 +758,7 @@ pub fn disk_info() -> Result<DiskInfo, Error> {
     {
         Ok(unsafe { get_disk_info() })
     }
-    #[cfg(any(target_os = "freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let mut di:DiskInfo = DiskInfo { total: 0, free: 0 };
         let res: i32 = unsafe { get_disk_info_bsd(&mut di) };
@@ -764,7 +768,7 @@ pub fn disk_info() -> Result<DiskInfo, Error> {
             _ => Err(Error::Unknown),
         }
     }
-    #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd")))]
+    #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd")))]
     {
         Err(Error::UnsupportedSystem)
     }
@@ -811,7 +815,7 @@ pub fn boottime() -> Result<timeval, Error> {
             tv_usec: secs.next().unwrap()? as libc::suseconds_t,
         })
     }
-    #[cfg(any(target_os = "macos", target_os="freebsd", target_os = "openbsd"))]
+    #[cfg(any(target_os = "macos", target_os="freebsd", target_os = "openbsd", target_os = "netbsd"))]
     {
         let mut bt = timeval {
             tv_sec: 0,